Refactor: Added more packages
Made a new package named Bonuses, which contains Armor, headgear, footgear, items, lefthand, rigthand and OneTimeUse.
A package with commands: Command, CommandWord, CommandWords and Parser.
