+++ WorldofZuul/src/Commands/Command.java
@@ -16,7 +16,7 @@
  * @author Michael Kolling and David J. Barnes
  * @version 2006.03.30
  */
-package worldofzuul;
+package Commands;
 
 public class Command {
