+++ WorldofZuul/src/Commands/CommandWord.java
@@ -1,4 +1,4 @@
-package worldofzuul;
+package Commands;
 
 public enum CommandWord {
     GO("go"), QUIT("quit"), HELP("help"), UNKNOWN("?"), FIGHT("fight"), FLEE("flee"), LOOT("loot"), /*ROLL("roll"),*/ CHARACTER("character");
