+++ WorldofZuul/src/Commands/CommandWords.java
@@ -1,5 +1,6 @@
-package worldofzuul;
+package Commands;
 
+import Commands.CommandWord;
 import java.util.HashMap;
 
 public class CommandWords {
