+++ WorldofZuul/src/Commands/Parser.java
@@ -1,5 +1,7 @@
-package worldofzuul;
+package Commands;
 
+import Commands.CommandWords;
+import Commands.Command;
 import java.util.Scanner;
 import java.util.StringTokenizer;
