+++ WorldofZuul/src/worldofzuul/Room.java
@@ -1,5 +1,6 @@
 package worldofzuul;
 
+import Bonuses.Item;
 import java.util.Set;
 import java.util.HashMap;
 import java.util.Iterator;
